Tests: Properly wait for the server to start up
I really don't like how this is done, but due to a super strange behaviour, this is the best solution I can think of...

On the first test that runs (not specifically the 404 test, but the first one that runs; I checked by running the tests in a different order) it combines the first and second stdout print into a single stream event, instead of two.
In all the subsequent tests, it properly separates the two, however...
So I cannot just neatly await the second event, because it sometimes just doesn't get fired.